+++ pywbem/cim_obj.py
@@ -2529,7 +2529,7 @@ class CIMProperty(_CIMComparisonMixin):
                 val_ = value_
             _mof = mofstr(val_, indent=indent)
         elif self.type == 'datetime':
-            _mof = '"%s' % mofstr(val_, indent=indent)
+            _mof = '"%s"' % str(value_)
         else:
             _mof = str(value_)
         return _mof